+++ src/org/javateerz/ParkViewProtector/VisualFX.java
@@ -9,7 +9,7 @@ public class VisualFX extends Movable implements Serializable
 	protected String name;
 	int time=0;
 	
-	public VisualFX(Game g, String name, int time, double x, double y)
+	public VisualFX(Game g, String name, double time, double x, double y)
 	{
 		super(g, x, y, 0);
 		this.name=name;
@@ -17,7 +17,7 @@ public class VisualFX extends Movable implements Serializable
 		updateSprite();
 	}
 	
-	public VisualFX(Game g, String name, int time)
+	public VisualFX(Game g, String name, double time)
 	{
 		super(g, 0, 0, 0);
 		this.name=name;
@@ -30,7 +30,7 @@ public class VisualFX extends Movable implements Serializable
 		this.sprite		= DataStore.INSTANCE.getSprite("fx/"+name+".png");
 	}
 	
-	public void setTime(int amount)
+	public void setTime(double amount)
 	{
 		time=ParkViewProtector.secsToFrames(amount);
 	}
